server: don't wait on idle connections when we have only 1 worker thread
When serving non multithread safe applications, a single worker thread is used, possibly with quite a few worker processes.

When a keep-alive request is dispatched, Puma checks for a new request coming in. This check makes any one worker thread stall if no data is available, up to FAST_TRACK_KA_TIMEOUT seconds (set to 200ms currently).

This has the effect of neglecting other requests that are ready, waiting to be picked up for processing.

This patch fixes the case for max_threads == 1, where waiting 200ms can greatly affect performance. It probably affects it too with more threads, but further benchmarks are needed to consider when it pays to wait up to 200ms.