server: queue requests if the option is enabled
In PR #1278 this was changed to avoid using the "queue_requests" params,
so workers only accept new connections if they have a free thread. This
commit reverts that.

I guess that change was beneficial for the most typical deployment
scenarios, but it has a noticeable performance penalty in the case of
Apisonator given that it is single threaded. Even with a moderate number
of connections there are lots of time-outs.